Station Filter Selection Screen Implementation
- Filter implemented

+++ charger_app/charger_app/Modules/StationFilter/FilterSettings/FilterSettings.swift
@@ -10,6 +10,6 @@ import Foundation
 class FilterSettings {
     var chargeType: [ChargeType]? = [ChargeType]()
     var socketTpye: [SocketType]? = [SocketType]()
-    var distance: Int?
+    var distance: Double?
     var serviceType: [ServiceType]? = [ServiceType]()
 }
